去除编译告警
git-svn-id: http://192.168.20.11/sequoiadb/trunk@21376 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/dps/dpsTransLockDef.cpp
@@ -123,7 +123,7 @@ namespace engine
 
    BOOLEAN dpsTransLockId::isValid() const
    {
-      if ( ~0 == _logicCSID )
+      if ( (UINT32)~0 == _logicCSID )
       {
          return FALSE ;
       }
